Added _read_mp4box_stem_titles function
Basically, for stems which have been created by the NI creator have the titles of the streams in the `udta` metadata.

I added the functionality to try to read the stream titles if MP4Box is available, else fall back to original default numbered streams.

This is useful if you want to extract wav files using the stem2files, and instead of getting filenames like "stem_0.wav", you can get "Drums.wav" or "Vox.wav" if this information is available in the stem files.

The code is taken almost verbatim from the `StemMetadataViewer` class in https://github.com/axeldelafosse/stemgen/blob/master/ni-stem/_internal.py